Stop tracking changes to `.env` after adding Dark Sky API Key
I added the Dark Sky API Key to `.env`, so that I can run the webserver
from my local machine. I don't want the API Key tracked in git, but I
don't want to remove the file from the repository, since it is a useful
placeholder when I clone the repo elsewhere.

By running the command:

```
$> git update-index --assume-unchanged .env
```

The file still exists in my local repository, but the changes to it will
be ignored.